+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix legacy `interviewId` parameter in `activeInterview` admin route (fixes [#1856](https://github.com/chairemobilite/evolution/issues/1856))
 - Redirect URLs returned by the `updateInterview` call do not log the participant out if it is internal, allowing in-survey redirections. (fixes [#1863](https://github.com/chairemobilite/evolution/issues/1863))
 - Fix spinner in admin's interview count by day graph when the count is 0 (fixes [#1870](https://github.com/chairemobilite/evolution/issues/1870))
+- Fix incomplete trip selection by selecting the first trip that has unsaved segments (fixes [#1868](https://github.com/chairemobilite/evolution/issues/1868))
 
 ### Security